# apps/api/openapi.yaml
# OpenAPI 2.0 (Swagger) spec consumed by GCP API Gateway.
#
# IMPORTANT: GCP API Gateway only accepts OpenAPI 2.0 — NOT the OpenAPI 3.0
# that FastAPI auto-generates at /openapi.json. Keep this file hand-edited
# in sync with apps/api/src/main.py until we add a 3.0 → 2.0 conversion step.
#
# `BACKEND_URL` is a placeholder substituted at deploy time by
# infra/api-gateway.ts with the Cloud Run service URL.

swagger: "2.0"
info:
    title: Text Digest API
    description: Authentication, Upload & Retrieval of E-Books, AI Q&A.
    version: "1.0.0"
schemes:
    - https
produces:
    - application/json
paths:
    /health:
        get:
            summary: Health check
            operationId: health
            x-google-backend:
                address: BACKEND_URL/health
            responses:
                "200":
                    description: API is healthy.
                    schema:
                        type: object
                        properties:
                            status:
                                type: string
                                example: ok
    /titles:
        get:
            summary: List the caller's titles
            operationId: listTitles
            x-google-backend:
                address: BACKEND_URL/titles
            responses:
                "200":
                    description: Array of title metadata.
        post:
            summary: Upload a PDF; processing happens async (publish-and-forget)
            operationId: uploadTitle
            x-google-backend:
                address: BACKEND_URL/titles
            responses:
                "202":
                    description: Accepted; returns taskId. Frontend subscribes to SSE progress directly against the Cloud Run URL (bypassing this gateway, which hard-caps backend deadline at 600s and buffers streamed responses).
    /titles/{title_id}:
        parameters:
            - name: title_id
              in: path
              required: true
              type: string
        get:
            summary: Get a single title with signed cover/markdown URLs and TOC
            operationId: getTitle
            x-google-backend:
                address: BACKEND_URL/titles/{title_id}
            responses:
                "200":
                    description: Title.
                "404":
                    description: Not found.
        delete:
            summary: Delete a title (Firestore + GCS, including chunks)
            operationId: deleteTitle
            x-google-backend:
                address: BACKEND_URL/titles/{title_id}
            responses:
                "204":
                    description: Deleted.